Migrate to Real/Simulated data options and protect proprietary data from non-ligo users
Summary: So that we can have gwcloud work with both proprietary and open/simulated data for both ligo and non-ligo users, we need to add checks for non-ligo users so that they can not every launch a job with proprietary data. Checks are also added to make sure that users can not view proprietary jobs or results if they are not a ligo user.

Test Plan:
Requires D436. Creating a new user who is not a ligo user in the database will result in only GWOSC channels being provided for real data. When detector cards are enabled, the channel select will be disabled. For a ligo user, they may choose to use any channel, and the channel select will be enabled.

Users who are non-ligo users should not be able to view any proprietary data job, or view any result files from proprietary jobs.

Reviewers: mr-asher, TomReichardt

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D437