Add preferred label to jobs
Summary: To help identify notable production run jobs, a new "Preferred" label needs to be added. This label should also be protected, such that a normal user may not add or delete this label from a job.

Test Plan: After patching this locally and applying migrations, it should not be possible to add the preferred lable to an existing job. If you manually apply the label using the django console, the preferred label should be visible to that job, and you should then not be able to remove the label via the webapp.

Reviewers: TomReichardt, jesmigel

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D519