Fix filesize overflow in graphene
Summary:
Files with sizes bigger than signed 32 bit numbers (2G) resulted in a graphene Int parsing error. To fix this we should use Decimal which is a virtually unlimited type.

Fix filesize overflow in graphene

Test Plan: I tested this locally by using a 110 Petabyte number in a Decimal, and the correct filesize was displayed in the UI without an error.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D531