Rename and migrate bilby django app to bilbyui
Summary:
As we move to using the bilby and bilby_pipe tools, we can no longer have a bilby django app as "bilby" becomes ambiguous and the incorrect module is loaded causing runtime errors. To fix that we need to rename the bilby django app to bilbyui to avoid module name conflicts.

The steps taken in this diff are as follows:-
* Rename the bilby app folder to bilbyui
* Update migrations to correctly generate the old "bilby_xxx" table names in the database
* Update bilby module references within the django application to correctly use bilbyui instead
* Add a new migration to rename the `bilby_xxx` tables to `bilbyui_xxx` table names.
* Introduce a new special python migration in the initial migration which detects any old bilby migrations applied in the migration database, and renames them to the bilbyui app.
  - If this occurs, it will then force commit those changes, then rerun the migration step and return.

Test Plan: Patching this locally, applying migrations, and running tests should all succeed as expected. Running the applicaton locally should also continue to run as expected. The CI will confirm that the migrations apply correctly to an empty database.

Reviewers: ssaleheen, TomReichardt

Reviewed By: ssaleheen, TomReichardt

Differential Revision: https://phab.adacs.org.au/D532