Add INI file parsing to key/value pairs
Summary:
So that we can easily filter for information about the job, we need to parse the ini file so we can easily get, for example, `trigger-time` for a particular job.

This diff adds the key/value pair model, and a migration which parses all jobs with ini_strings.

Test Plan:
Patching this locally and migrating should result in any local jobs with ini_strings being parsed and each key=value should end up verbatim in the IniKeyValue model.

Any attributes missing from an ini file will be added to the k/v model with the default value as specified by bilby_pipe.

Running tests should also pass.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D534