Implement ini parsing for the UI and update UI to work with new schema
Summary:
Now that we have ini parsing for submitting jobs, and we've removed the old parameter models; we need to parse the ini_string to return job information for consumption by the UI. The UI also needed some work to work with the new schema.

A bunch of end to end testing of jobs from job submission to job retrieval has been added.

Test Plan: Patching this locally should allow you to view an existing job in the UI, and duplicating/submitting jobs should also work as expected. There are some cases where information is missing - but this is expected, and newly submitted jobs will not have missing data. Mostly the missing data is around sampler arguments - since they were previously not saved at all. Deactivated detectors do not store any details in the ini file.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D538