Fix MySQL unique constraint error
Summary:
As it turns out, MySQL doesn't like unique constraints on indexes of different types. It's actually invalid. This diff fixes the adds a new migration to correctly create the IniKeyValue table and upates the models.

See:-
* https://stackoverflow.com/questions/24106756/create-table-fails-with-foreign-key-constraint-is-incorrectly-formed
* https://stackoverflow.com/questions/8434518/mysql-foreign-key-constraint-is-incorrectly-formed-error

Test Plan: All tests should pass without issue. Eventually when this is landed, the db search service should be able to run tests successfully, and we should be able to deploy gwcloud in to production.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D540