Fix bundle to work correctly with bilby ui
Summary:
The bundle was incorrectly assuming that the ini_string would be magically captured by the job client and sent as a raw ini string to the submit function. Unfortunately little elves in the code didn't work like that at all, so we need to correctly parse the json sent frome the gwcloud UI.

Additionally the `working_directory` function is called directly on occasion from the job client. However it's called with two arguments, even though the second argument is not used. Therefore we need to add args/kwargs in to handle this case.

Test Plan: Tests are updated and pass.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D550