Override ini label attribute when submitting ini file
Summary:
Previously when submitting an ini file from the API, the internal label attribute of the ini file was not overridden to match the name provided by the job details sent from the API. This resulted in jobs running remotely with incorrect job names, and output results that didn't match the BilbyJob name.

This diff fixes that, and provides and tests correct ini file label attribute overrides.

Test Plan: A new test has been added which checks that the label attribute in submitted ini files is correctly overridden.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D573