Add uploaded file download
Summary: This diff adds code that allows files uploaded via job upload to be downloaded. The download parameters and logic matches the logic from the job controller, so when downloading files from this endpoint - we can use the same logic and flow as we use for the job controller. The caveat here being that for uploaded jobs, file download tokens and file download id's are identical. Requesting file download ids is a noop for an uploaded job. This is intentional to remove or reduce the overhead of having to modify API consumers.

Test Plan: All code has been tested. However patching this locally should now allow you to use the API to upload a job, list the job files, and then download job files.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D586