Superclass scheduler in bundle
Summary:
So that many schedulers can be supported, we need to move the current slurm scheduler to a class and create a superclass for other schedulers that can be inherited from.

This diff creates a Scheduler superclass and moves the slurm submission code to an inherited class.

Test Plan: Tests have been updated for testing job submission and they pass as expected.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D636