Add submission cluster to django
Summary: So that a user can choose which cluster to submit a job to, we need to add a cluster parameter to the graphql query and then check that the cluster is valid before submitting the job.

Test Plan: New tests have been added which test ini and non-ini job submission using the graphql api when using various permutations of clusters.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D637