Integrate bundle coverage and tests with Jenkins
Summary:
So that we can improve and track code quality and coverage in the bundle, we need to integrate it with jenkins.

This diff adds a few packages to the bundle conda environment that can be used for generating cobertura coverage xml, and junit xml for consumption by jenkins.

Test Plan: Jenkins passes all tests and reports coverage including bundle coverage.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D667