Add Condor scheduler
Summary:
This diff adds the Condor scheduler. It includes a condor submit function and a condor status function which parses the job submit node log to determine the job status.

Other general refactoring and comprehensive submit/status tests are added.

Test Plan:
Thorough tests are added for the most important parts of the bundle, submit and status.

Lint errors in settings.py are OK.

Reviewers: TomReichardt, mr-asher

Reviewed By: TomReichardt, mr-asher

Differential Revision: https://phab.adacs.org.au/D663