Run pip-compile to handle new bilby_pipe dep changes
Summary: So that the CI doesn't bork, and so that other venvs stay sane with the new bilby_pipe dependency version changes, we need to rerun pip-compile. This is a stop gap until bilby_pipe 1.0.5 is formally released (soon).

Test Plan: Run `pip install -r requirements.txt`, if you run in to errors, it might be possible you need to delete your current venv and create a new one then install the packages.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D750