Bilby supporting file parsing and uploading
So that we can handle supporting file uploads for Bilby jobs, we need to be able to parse the ini file for supporting files, return the supporting files to the user if there are any, then wait for all the supporting files to be uploaded and submit the job. Jobs with no supporting files should be submitted immediately without this intermediate step.

The result from parsing the ini file for supporting files should be an array of dictionaries containing the supporting file upload token and the path to the file as per the ini file. This is so that the client doesn't need to have bilby_pipe installed locally to parse the ini file. The client should then upload each supporting file using the local path from the parsed ini file and the upload token.

Merged by: Asher <jasherleslie@gmail.com>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_bilby/-/merge_requests/39