Bundle: Work with new parallel final Bilby job steps
During the update to the latest stable bilby_pipe version, the format of dependent jobs changed such that either of two job stages could finish last. A new job stage "final_result" was added that can run concurrently with "plot". This caused an issue in the condor scheduler where we incorrectly assumed the job was still running even though it was not. The new condor status code assumes that the last stage is plotting, and waits for it to finish - however if an optional final result step is started before or after the plotting stage then that is awaited before the job is considered finished as well.

Approved-by: TomReichardt <thomas.reichardt@students.mq.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_bilby/-/merge_requests/48