Refactor BilbyJob schema user and job service requests
The job list pages are starting to load very slowly due to the number of jobs that are making requests in the background to get user and job status information.

Instead we should make sure that we make only one request for each query rather than for each object in the query. The external API's support fetching information for multiple jobs or users per request.

In this MR the full list of users and job details is fetched in the query set function, and are then passed down via the query context to the relevant resolvers for each object. A simple get is then made to fetch the object (or None) from the cached user or job controller dicts.

Approved-by: Shibli Saleheen <ssaleheen@swin.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_bilby/-/merge_requests/74