[backend] Fix relay infinite scroll issue on Public Jobs page
It turns out that passing a variables dictionary vs providing variable values directly in the query generate different results. The previous tests relied on the latter, while the frontend relied on the former. Now that the tests have been updated to use a variable dict, we've uncovered a strange occurrence with graphene-django where it slices the array after the resolver has returned the objects. So we need to make sure that the resolver returns an array with the expected number of elements.

Approved-by: Asher <jasherleslie@gmail.com>
Approved-by: TomReichardt <thomas.reichardt@students.mq.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_bilby/-/merge_requests/89