Fix special symbols in event json
We were so worried about what would happen if there was a symbol (:.~)
which can't be in a bilbyJob name present in a h5 key that we didn't
stop to think about what would happen if there was one present in the
event json.

What happens, is that it breaks everything because `fix_job_name(k)` is
not the same as `k` when it comes to indexing a json file.